✏️rename `Search` folder to `GlobalSearch`
