+++ src/pages/GlobalSearch.tsx
@@ -4,8 +4,8 @@ import { View, StyleSheet } from "react-native"
 import { MainStackScreen } from "navigation/NavigationTypes"
 import { ListPage } from "components/PageLayout"
 import { PoliSearchBar } from "components/Home"
-import { SearchTagFilter } from "components/Search/SearchTagFilter"
-import { SearchTile } from "components/Search/SearchTile"
+import { SearchTagFilter } from "components/GlobalSearch/SearchTagFilter"
+import { SearchTile } from "components/GlobalSearch/SearchTile"
 import { SearchTag, SEARCH_TAGS, SEARCH_TAG_TO_LABEL } from "utils/search"
 
 export const GlobalSearch: MainStackScreen<"GlobalSearch"> = () => {
